xsk: XDP socket ring LKMM litmus tests
The XDP socket single-produce/single-consumer ring was initially based
on barriers (smp_{r,w,}mb(). For effeciency and ease of use, we'd like
to move the rings to an acquire/release semantic instead, but making
sure that we don't break userland.

These LKMM litmus tests aid that transition. The litmus tests is based
on a one entry ring-buffer, and there are two scenarios. The ring is
full, i.e. the producer has written an entry, so the consumer has to
go first. The ring is empty, i.e. the producer has to go first.

There is one test for each permutation: barrier only, acqrel only,
acqrel+barrier, barrier+acqrel. In total eight scripts.

Signed-off-by: Björn Töpel <bjorn.topel@intel.com>